chat-widget: Move source code to apps/
Change-type: patch
