+++ Makefile
@@ -261,18 +261,18 @@ endif
 
 ifeq ($(COVERAGE),1)
 build-chat-widget:
-	rm -rf $(NYC_TMP_DIR) && mkdir -p $(NYC_TMP_DIR)
-	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/ui $(NYC_TMP_DIR)/ui
-	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/sdk $(NYC_TMP_DIR)/sdk
-	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/ui-components $(NYC_TMP_DIR)/ui-components
-	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/chat-widget $(NYC_TMP_DIR)/chat-widget
-	NODE_ENV=test UI_DIRECTORY="./$(NYC_TMP_DIR)/chat-widget" \
+	rm -rf $(NYC_TMP_DIR) && mkdir -p $(NYC_TMP_DIR)/{lib,apps}
+	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/ui $(NYC_TMP_DIR)/lib/ui
+	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/sdk $(NYC_TMP_DIR)/lib/sdk
+	./node_modules/.bin/nyc instrument $(NYC_OPTS) lib/ui-components $(NYC_TMP_DIR)/lib/ui-components
+	./node_modules/.bin/nyc instrument $(NYC_OPTS) apps/chat-widget $(NYC_TMP_DIR)/apps/chat-widget
+	NODE_ENV=test \
 		SENTRY_DSN_UI=$(SENTRY_DSN_UI) API_URL=$(SERVER_HOST):$(SERVER_PORT) \
 		CHAT_WIDGET_JELLYFISH_TOKEN=$(CHAT_WIDGET_JELLYFISH_TOKEN) \
 		./node_modules/.bin/webpack --config=./apps/chat-widget/webpack.config.js
 else
 build-chat-widget:
-	UI_DIRECTORY="./lib/chat-widget" SENTRY_DSN_UI=$(SENTRY_DSN_UI) API_URL=$(SERVER_HOST):$(SERVER_PORT) \
+	SENTRY_DSN_UI=$(SENTRY_DSN_UI) API_URL=$(SERVER_HOST):$(SERVER_PORT) \
 	CHAT_WIDGET_JELLYFISH_TOKEN=$(CHAT_WIDGET_JELLYFISH_TOKEN) \
 		./node_modules/.bin/webpack --config=./apps/chat-widget/webpack.config.js
 endif
