+++ apps/chat-widget/components/ErrorMessage.jsx
