+++ apps/chat-widget/components/Header.jsx
