+++ apps/chat-widget/components/Layout.jsx
