+++ apps/chat-widget/routes/FullThreadListRoute.jsx
@@ -6,10 +6,10 @@ import {
 	Box,
 	Flex
 } from 'rendition'
-import CardChatSummary from '../../ui-components/CardChatSummary'
+import CardChatSummary from '../../../lib/ui-components/CardChatSummary'
 import {
 	InfiniteList
-} from '../../ui-components/InfiniteList'
+} from '../../../lib/ui-components/InfiniteList'
 import {
 	Loader
 } from '../components/Loader'
