+++ apps/chat-widget/store/actionTypes.js
